move the member profile tour
